Check for existing billed rows. Add success/error
Output for users will be clearer when a charge fails to get posted
to the Autotask api. The omission of -Force on the Addition of the
charge row to the AzDataTable enables duplicate charges from being
entered. Adding the Autotask charge AFTER that call makes the
exeception prevent the API call to Autotask.